Major Update
Cleaned up a large number of the games, up to 5 minigames. Added win triggers for all the mouse players, and added the system to change the minigame with smaller systems in place for overall game management and scoring. Most games still lack timer triggers for keyboard players (will be located in one script) - and balance is still untouched.